extracting ProcessIcon component
