+++ src/lib/components/process/ProcessTable.svelte
@@ -1,7 +1,12 @@
 <script lang="ts">
+  import {
+    faThumbtack,
+    faInfoCircle,
+    faXmark,
+  } from "@fortawesome/free-solid-svg-icons";
+  import Fa from "svelte-fa";
   import type { Process, Column } from "$lib/types";
-  import TableHeader from "./TableHeader.svelte";
-  import ProcessRow from "./ProcessRow.svelte";
+  import ProcessIcon from "./ProcessIcon.svelte";
 
   export let processes: Process[];
   export let columns: Column[];
@@ -13,23 +18,81 @@
   export let onTogglePin: (command: string) => void;
   export let onShowDetails: (process: Process) => void;
   export let onKillProcess: (process: Process) => void;
+
+  function getSortIndicator(field: keyof Process) {
+    if (sortConfig.field !== field) return "↕";
+    return sortConfig.direction === "asc" ? "↑" : "↓";
+  }
 </script>
 
 <div class="table-container">
   <table>
-    <TableHeader {columns} {sortConfig} {onToggleSort} />
+    <thead>
+      <tr>
+        {#each columns.filter((col) => col.visible) as column}
+          <th class="sortable" on:click={() => onToggleSort(column.id)}>
+            <div class="th-content">
+              {column.label}
+              <span
+                class="sort-indicator"
+                class:active={sortConfig.field === column.id}
+              >
+                {getSortIndicator(column.id)}
+              </span>
+            </div>
+          </th>
+        {/each}
+        <th>Actions</th>
+      </tr>
+    </thead>
     <tbody>
       {#each processes as process (process.pid)}
-        <ProcessRow
-          {process}
-          {columns}
-          isPinned={pinnedProcesses.has(process.command)}
-          isHighUsage={process.cpu_usage > 50 ||
+        <tr
+          class:high-usage={process.cpu_usage > 50 ||
             process.memory_usage / (systemStats?.memory_total || 0) > 0.1}
-          {onTogglePin}
-          {onShowDetails}
-          {onKillProcess}
-        />
+          class:pinned={pinnedProcesses.has(process.command)}
+        >
+          {#each columns.filter((col) => col.visible) as column}
+            <td class="truncate">
+              {#if column.id === "name"}
+                <div class="name-cell">
+                  <ProcessIcon processName={process.name} />
+                  <span class="process-name">{process.name}</span>
+                </div>
+              {:else if column.format}
+                {@html column.format(process[column.id])}
+              {:else}
+                {process[column.id]}
+              {/if}
+            </td>
+          {/each}
+          <td class="col-actions">
+            <div class="action-buttons">
+              <button
+                class="btn-action pin-btn"
+                class:pinned={pinnedProcesses.has(process.command)}
+                on:click={() => onTogglePin(process.command)}
+                title={pinnedProcesses.has(process.command) ? "Unpin" : "Pin"}
+              >
+                <Fa icon={faThumbtack} />
+              </button>
+              <button
+                class="btn-action info-btn"
+                on:click={() => onShowDetails(process)}
+                title="Show Details"
+              >
+                <Fa icon={faInfoCircle} />
+              </button>
+              <button
+                class="btn-action kill-btn"
+                on:click={() => onKillProcess(process)}
+                title="End Process"
+              >
+                <Fa icon={faXmark} />
+              </button>
+            </div>
+          </td>
+        </tr>
       {/each}
     </tbody>
   </table>
@@ -82,5 +145,212 @@
     table-layout: fixed;
     border-collapse: collapse;
     font-size: 13px;
+  }
+
+  th {
+    position: sticky;
+    top: 0;
+    background: var(--mantle);
+    text-align: left;
+    padding: 8px 12px;
+    font-weight: 500;
+    color: var(--subtext0);
+    border-bottom: 1px solid var(--surface0);
+    transition: background-color 0.2s ease;
+    z-index: 3;
+  }
+
+  td {
+    padding: 6px 12px;
+    border-bottom: 1px solid var(--surface0);
+    color: var(--text);
+    z-index: 1;
+  }
+
+  tr:hover {
+    background-color: var(--surface0);
+  }
+
+  .truncate {
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    max-width: 0;
+  }
+
+  .sortable {
+    cursor: pointer;
+    user-select: none;
+  }
+
+  .th-content {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+  }
+
+  .sort-indicator {
+    color: var(--overlay0);
+    font-size: 12px;
+    opacity: 0.5;
+    transition: all 0.2s ease;
+  }
+
+  .sort-indicator.active {
+    color: var(--blue);
+    opacity: 1;
+  }
+
+  .sortable:hover .sort-indicator {
+    opacity: 1;
+  }
+
+  .high-usage {
+    background-color: color-mix(in srgb, var(--red) 10%, transparent);
+  }
+
+  .high-usage:hover {
+    background-color: color-mix(in srgb, var(--red) 15%, transparent);
+  }
+
+  tr.pinned {
+    background-color: color-mix(in srgb, var(--blue) 10%, transparent);
+  }
+
+  tr.pinned:hover {
+    background-color: color-mix(in srgb, var(--blue) 15%, transparent);
+  }
+
+  th:last-child {
+    width: 120px;
+    min-width: 120px;
+    max-width: 120px;
+  }
+
+  td:last-child {
+    width: 120px;
+    min-width: 120px;
+    max-width: 120px;
+    padding: 6px 8px;
+  }
+
+  .col-actions {
+    position: sticky;
+    right: 0;
+    z-index: 2;
+    background: var(--base);
+    border-left: 1px solid var(--surface0);
+    width: 120px;
+  }
+
+  .action-buttons {
+    display: flex;
+    gap: 4px;
+    align-items: center;
+    justify-content: center;
+  }
+
+  .btn-action {
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    border: none;
+    cursor: pointer;
+    background: transparent;
+    border-radius: 6px;
+    width: 28px;
+    height: 28px;
+    transition: all 0.2s ease;
+    position: relative;
+    overflow: hidden;
+  }
+
+  .btn-action::before {
+    content: "";
+    position: absolute;
+    inset: 0;
+    opacity: 0.1;
+    transition: opacity 0.2s ease;
+  }
+
+  .btn-action:hover::before {
+    opacity: 0.15;
+  }
+
+  .pin-btn {
+    color: var(--sapphire);
+  }
+
+  .pin-btn::before {
+    background: var(--sapphire);
+  }
+
+  .pin-btn.pinned {
+    color: var(--blue);
+    transform: rotate(45deg);
+  }
+
+  .pin-btn.pinned::before {
+    background: var(--blue);
+    opacity: 0.15;
+  }
+
+  .info-btn {
+    color: var(--lavender);
+  }
+
+  .info-btn::before {
+    background: var(--lavender);
+  }
+
+  .kill-btn {
+    color: var(--red);
+    border: 1px solid color-mix(in srgb, var(--red) 30%, transparent);
+  }
+
+  .kill-btn:hover {
+    color: var(--base);
+    background: var(--red);
+  }
+
+  .kill-btn:hover::before {
+    opacity: 1;
+  }
+
+  .btn-action:hover {
+    box-shadow: 0 0 12px color-mix(in srgb, currentColor 20%, transparent);
+  }
+
+  .btn-action:active {
+    transform: translateY(1px);
+  }
+
+  .pin-btn.pinned:active {
+    transform: rotate(45deg) translateY(1px);
+  }
+
+  .btn-action:focus {
+    outline: none;
+    box-shadow: 0 0 0 2px color-mix(in srgb, currentColor 30%, transparent);
+  }
+
+  .btn-action:disabled {
+    opacity: 0.5;
+    cursor: not-allowed;
+  }
+
+  .btn-action:disabled:hover {
+    transform: none;
+    box-shadow: none;
+  }
+
+  .btn-action:disabled::before {
+    display: none;
+  }
+
+  .name-cell {
+    display: flex;
+    align-items: center;
+    gap: 8px;
   }
 </style>
